# app/mgmt/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "此文档由gin-swagger自动生成",
        "title": "xxim-server HTTP API 文档",
        "contact": {
            "name": "showurl"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:6799",
    "basePath": "/api",
    "paths": {
...
        "/ms/add/notice": {
            "post": {
                "description": "使用此接口新增app公告",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app公告管理"
                ],
                "summary": "新增app公告",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.AddAppMgmtNoticeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.AddAppMgmtNoticeResp"
                        }
                    }
                }
            }
        },
...
        "/ms/delete/notice": {
            "post": {
                "description": "使用此接口删除app公告",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app公告管理"
                ],
                "summary": "删除app公告",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteAppMgmtNoticeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteAppMgmtNoticeResp"
                        }
                    }
                }
            }
        },
...
        "/ms/get/notice/detail": {
            "post": {
                "description": "使用此接口获取app公告详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app公告管理"
                ],
                "summary": "获取app公告详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAppMgmtNoticeDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.GetAppMgmtNoticeDetailResp"
                        }
                    }
                }
            }
        },
        "/ms/get/notice/list/all": {
            "post": {
                "description": "使用此接口获取全部app公告列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app公告管理"
                ],
                "summary": "获取全部app公告列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllAppMgmtNoticeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllAppMgmtNoticeResp"
                        }
                    }
                }
            }
        },
...
        "/ms/get/version/detail": {
            "post": {
                "description": "使用此接口获取app版本详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app版本管理"
                ],
                "summary": "获取app版本详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAppMgmtVersionDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.GetAppMgmtVersionDetailResp"
                        }
                    }
                }
            }
        },
        "/ms/get/version/list/all": {
            "post": {
                "description": "使用此接口获取全部app版本列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app版本管理"
                ],
                "summary": "获取全部app版本列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllAppMgmtVersionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.GetAllAppMgmtVersionResp"
                        }
                    }
                }
            }
        },
...
        "/ms/update/notice": {
            "post": {
                "description": "使用此接口更新app公告",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app公告管理"
                ],
                "summary": "更新app公告",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateAppMgmtNoticeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应数据",
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateAppMgmtNoticeResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.AddAppMgmtNoticeReq": {
            "type": "object",
            "properties": {
                "appMgmtNotice": {
                    "$ref": "#/definitions/pb.AppMgmtNotice"
                },
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                }
            }
        },
        "pb.AddAppMgmtNoticeResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.AppMgmtNotice": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "createdAtStr": {
                    "type": "string"
                },
                "endTime": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "isEnable": {
                    "type": "boolean"
                },
                "platform": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "startTime": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "pb.AppMgmtVersion": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "createdAtStr": {
                    "type": "string"
                },
                "downloadUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "pb.CommonReq": {
            "type": "object",
            "properties": {
                "appVersion": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "deviceId": {
                    "type": "string"
                },
                "deviceModel": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "osVersion": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "userAgent": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "pb.CommonResp": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/pb.CommonResp_Code"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "pb.CommonResp_Code": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                7,
                8,
                9,
                10
            ],
            "x-enum-comments": {
                "CommonResp_AlertError": "alert 错误 只有一个确定按钮",
                "CommonResp_AuthError": "鉴权错误 // 应该退出登录",
                "CommonResp_ForbiddenError": "无权限",
                "CommonResp_InternalError": "内部错误",
                "CommonResp_NeedSecondPasswordError": "需要输入二级密码/设置二级密码",
                "CommonResp_RequestError": "请求错误",
                "CommonResp_RetryError": "alert 错误 有一个取消按钮 和一个重试按钮",
                "CommonResp_ToastError": "toast 错误 只有 message",
                "CommonResp_UnknownError": "未知 error"
            },
            "x-enum-varnames": [
                "CommonResp_Success",
                "CommonResp_UnknownError",
                "CommonResp_InternalError",
                "CommonResp_RequestError",
                "CommonResp_AuthError",
                "CommonResp_ToastError",
                "CommonResp_AlertError",
                "CommonResp_RetryError",
                "CommonResp_ForbiddenError",
                "CommonResp_NeedSecondPasswordError"
            ]
        },
...
        "pb.DeleteAppMgmtNoticeReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "pb.DeleteAppMgmtNoticeResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.GetAllAppMgmtNoticeReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "page": {
                    "$ref": "#/definitions/pb.Page"
                }
            }
        },
        "pb.GetAllAppMgmtNoticeResp": {
            "type": "object",
            "properties": {
                "appMgmtNotices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.AppMgmtNotice"
                    }
                },
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "pb.GetAllAppMgmtVersionReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "page": {
                    "$ref": "#/definitions/pb.Page"
                }
            }
        },
        "pb.GetAllAppMgmtVersionResp": {
            "type": "object",
            "properties": {
                "appMgmtVersions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.AppMgmtVersion"
                    }
                },
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "pb.GetAppMgmtNoticeDetailReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "pb.GetAppMgmtNoticeDetailResp": {
            "type": "object",
            "properties": {
                "appMgmtNotice": {
                    "$ref": "#/definitions/pb.AppMgmtNotice"
                },
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.GetAppMgmtVersionDetailReq": {
            "type": "object",
            "properties": {
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "pb.GetAppMgmtVersionDetailResp": {
            "type": "object",
            "properties": {
                "appMgmtVersion": {
                    "$ref": "#/definitions/pb.AppMgmtVersion"
                },
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
        "pb.Page": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "pb.UpdateAppMgmtNoticeReq": {
            "type": "object",
            "properties": {
                "appMgmtNotice": {
                    "$ref": "#/definitions/pb.AppMgmtNotice"
                },
                "commonReq": {
                    "$ref": "#/definitions/pb.CommonReq"
                }
            }
        },
        "pb.UpdateAppMgmtNoticeResp": {
            "type": "object",
            "properties": {
                "commonResp": {
                    "$ref": "#/definitions/pb.CommonResp"
                }
            }
        },
...
    }
}